requirements.txt
python-dotenv
streamlit
faiss-cpu
bs4
groq
fastapi
wikipedia
google-api-python-client
langchain-google-community
langserve
langchain
langchain_core
langchain_community
langchain_openai
langchainhub
langchain-groq
langchain-objectbox
ipykernel
InstructorEmbedding
pathlib
yaml